refactor(core-api): move utils into shared controller
