+++ packages/core-api/src/services/index.ts
@@ -1,4 +1,3 @@
 import { ServerCache } from "./cache";
-import { transformerService } from "./transformer";
 
-export { ServerCache, transformerService };
+export { ServerCache };
